Break up SSH Interface Module
